Optimise Qemu Fix
Qemu appears to have a memory address conflict with the RefindPlus Title 
Banner's Pixel Data.

Until a proper solution is found, a memory leak has been permitted when 
ConsoleControl is absent (such as when Qemu is active). 

This commit limits the permitted leak to the Title Banner as opposed to 
all images as was the case previously (3ff6796)